Fixed MY_TYPE for RTAUDIO_SINT32 so that it's int32_t as `long` is 64-bit on some 64-bit systems
Closes https://github.com/thestk/rtaudio/issues/420

+++ tests/duplex.cpp
@@ -25,7 +25,7 @@ typedef signed short MY_TYPE;
 typedef S24 MY_TYPE;
 #define FORMAT RTAUDIO_SINT24
 
-typedef signed long MY_TYPE;
+typedef int32_t MY_TYPE;
 #define FORMAT RTAUDIO_SINT32
 
 typedef float MY_TYPE;
