+++ tests/playsaw.cpp
@@ -28,7 +28,7 @@ typedef S24 MY_TYPE;
 #define FORMAT RTAUDIO_SINT24
 #define SCALE  8388607.0
 
-typedef signed long MY_TYPE;
+typedef int32_t MY_TYPE;
 #define FORMAT RTAUDIO_SINT32
 #define SCALE  2147483647.0
